Consolidate the app into worker.ts; hand-rolled unzip; resilient app fan-out
Per review: the flake-dashboard app is now one giant worker.ts —
FlakeDashboardApp.processEvent ingests webhooks inline and wakes the
processor, FlakeDashboardProcessor reduces, render sits below; the
webhook shape is a plain zod schema in contract.ts next to its siblings
(no parser helper, bare safeParse); ingestion tests drive the real DO's
processEvent. fflate is out of production code: a ~55-line central-
directory zip reader on the runtime's own DecompressionStream (narrow
format surface: one producer, 5MB cap, no zip64; stored + deflate
entries both covered by tests) — fflate remains a devDependency for
building test-fixture zips, and both alwaysBundle entries are reverted.
The config template's app fan-out is now Promise.allSettled + aggregate
rethrow: every app sees every event even when a sibling throws, while
redelivery (at-least-once) is preserved for the app that failed.
