+++ packages/iterate/package.json
@@ -218,7 +218,6 @@
     "@orpc/server": "1.13.14",
     "@tanstack/react-query": "^5.101.2",
     "approx-string-match": "^2.0.0",
-    "fflate": "^0.8.3",
     "mdast-util-from-markdown": "^2.0.3",
     "mdast-util-gfm": "^3.1.0",
     "micromark-extension-gfm": "^3.0.0",
@@ -251,6 +250,7 @@
     "@types/react-dom": "^19.2.3",
     "@types/ws": "^8.18.1",
     "esbuild": "^0.27.4",
+    "fflate": "^0.8.3",
     "jsdom": "27.4.0",
     "miniflare": "catalog:cloudflare",
     "octokit": "^5.0.5",
